Move out many more tests from StripeTest
